fix(guest-access): mark copied passwords as sensitive
CreatedConversation already flagged the clipboard entry with
EXTRA_IS_SENSITIVE and confirmed the copy with a toast, but the two copy
actions this branch adds did neither, so on Android 13 and later the
password appeared in the clipboard preview and nothing told the user it
had been copied.

Lift that private helper into ClipboardUtils and use it at all three
sites.
